requirements.txt
crewai
langchain
crewai_tools
python-dotenv
langchain-huggingface
langchain-community
langchain-groq
youtube-transcript-api
litellm